refactor(dialog): move dialog animations from global css to scoped module
Refactored the DialogContent component to use scoped CSS modules instead of global animations.
Moved DialogContent to its own directory with a dedicated SCSS module file.
Removed unused global animation variables from animations.css.
Exported DialogTitle and DialogDescription from radix-ui for better accessibility.
