+++ src/assets/css/theme/animations.css
@@ -1,11 +1,8 @@
 @theme static {
  /* Animations */
- --animate-fade-in: fadein 0.3s ease-out forwards;
- --animate-fade-out: fadeout 0.2s ease-in forwards;
+
  --animate-language-close: language-close 0.15s ease-in-out;
  --animate-language-open: language-open 0.15s ease-in-out;
- --animate-scale-in: scalein 0.3s ease-out forwards;
- --animate-scale-out: scaleout 0.2s ease-in forwards;
 
  /* Keyframes */
  @keyframes fadeIn {
@@ -45,30 +42,6 @@
 
   100% {
    height: 160px;
-  }
- }
-
- @keyframes scaleIn {
-  from {
-   transform: scale(0.95) translate(-50%, -50%);
-   opacity: 0;
-  }
-
-  to {
-   transform: scale(1) translate(-50%, -50%);
-   opacity: 1;
-  }
- }
-
- @keyframes scaleOut {
-  from {
-   transform: scale(1) translate(-50%, -50%);
-   opacity: 1;
-  }
-
-  to {
-   transform: scale(0.95) translate(-50%, -50%);
-   opacity: 0;
   }
  }
 }
