+++ src/shared/Dialog/DialogContent/DialogContent.module.scss
@@ -1,0 +1,97 @@
+.overlay {
+ position: fixed;
+ inset: 0;
+ z-index: 0;
+ background-color: rgb(0 0 0 / 50%);
+
+ &[data-state="closed"] {
+  animation: fadeOut 0.2s ease-in forwards;
+ }
+
+ &[data-state="open"] {
+  animation: fadeIn 0.3s ease-out forwards;
+ }
+
+ @keyframes fadeIn {
+  0% {
+   opacity: 0;
+  }
+
+  100% {
+   opacity: 1;
+  }
+ }
+
+ @keyframes fadeOut {
+  0% {
+   opacity: 1;
+  }
+
+  100% {
+   opacity: 0;
+  }
+ }
+}
+
+.content {
+ position: fixed;
+ top: 50%;
+ left: 50%;
+ z-index: 1;
+ padding: var(--spacing-15px);
+ background-color: var(--color-white);
+ border-radius: var(--spacing-16px);
+ transform: translate(-50%, -50%);
+ transform: translate(-50%, -50%) scale(1);
+ transform-origin: top left;
+ will-change: transform;
+
+ &[data-state="closed"] {
+  animation: scaleOut 0.2s ease-in forwards;
+ }
+
+ &[data-state="open"] {
+  animation: scaleIn 0.3s ease-out forwards;
+ }
+
+ @media (width >= 768px) {
+  padding: var(--spacing-30px);
+  border-radius: var(--spacing-20px);
+ }
+
+ @keyframes scaleIn {
+  from {
+   transform: scale(0.95) translate(-50%, -50%);
+   opacity: 0;
+  }
+
+  to {
+   transform: scale(1) translate(-50%, -50%);
+   opacity: 1;
+  }
+ }
+
+ @keyframes scaleOut {
+  from {
+   transform: scale(1) translate(-50%, -50%);
+   opacity: 1;
+  }
+
+  to {
+   transform: scale(0.95) translate(-50%, -50%);
+   opacity: 0;
+  }
+ }
+}
+
+.close {
+ position: absolute;
+ top: var(--spacing-10px);
+ right: var(--spacing-10px);
+ cursor: pointer;
+
+ @media (width >= 768px) {
+  top: var(--spacing-20px);
+  right: var(--spacing-20px);
+ }
+}
