+++ src/shared/Dialog/DialogContent/DialogContent.tsx
@@ -11,7 +11,9 @@ import {
  Portal,
 } from "@radix-ui/react-dialog";
 
-import { DialogClose } from "./DialogClose";
+import { DialogClose } from "../DialogClose";
+
+import s from "./DialogContent.module.scss";
 
 export interface DialogContentProperties extends ContentProperties {
  children: ReactNode;
@@ -34,18 +36,10 @@ export const DialogContent: FC<DialogContentProperties> = (properties) => {
 
  return (
   <Portal {...portalProps}>
-   <Overlay
-    className="data-[state=closed]:animate-fadeOut data-[state=open]:animate-fadeIn fixed inset-0 bg-[rgba(0,0,0,0.50)]"
-    {...overlayProps}
-   />
-   <Content
-    className={cn(
-     "fixed top-1/2 left-1/2 -translate-x-1/2 -translate-y-1/2 bg-white mob:p-15px tablet:p-30px mob:rounded-2xl tablet:rounded-[1.25rem] transform-gpu data-[state=closed]:animate-scaleOut data-[state=open]:animate-scaleIn will-change-transform origin-top-left",
-     className,
-    )}
-    {...otherProperties}>
+   <Overlay className={s.overlay} {...overlayProps} />
+   <Content className={cn(s.content, className)} {...otherProperties}>
     {withCloseIcon ? (
-     <DialogClose className="mob:right-[10px] mob:top-[10px] tablet:right-5 tablet:top-5 absolute">
+     <DialogClose className={s.close}>
       <button type="button">
        <CloseIcon />
       </button>
